Rename inaccurately named storypoints.csv
